backend/requirements.txt
flask[async]
flask-cors
openai
python-dotenv
psycopg2-binary
pypdf
pydantic
duckduckgo-search
chromadb
volcengine
aiohttp
